requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.12.0
requests==2.31.0
boto3==1.34.34 